Add models and routes for budget, client, product, shipment, and drone management; enhance drone and warehouse routes with preload functionality

+++ internal/models/product.go
@@ -6,7 +6,7 @@ type Product struct {
 	BaseModel
 	Name        string    `gorm:"uniqueIndex;not null" json:"name"`
 	Description string    `gorm:"not null" json:"description"`
-	Price       int       `gorm:"not null" json:"price"`
+	Price       float64       `gorm:"not null" json:"price"`
 	ProviderID  uuid.UUID `gorm:"type:uuid;not null" json:"provider_id"`
 	Provider    *Provider `gorm:"foreignKey:ProviderID;references:ID" json:"provider"`
 }
